Adds highlighting to individual CI detail section
The individual correction instruction, which can be accessed from
the Market Place is now also highlighted.

I decided to leave the background untouched, as the context,
deletion and insertion blocks are already separated.

It could be that for big CIs with big blocks it would be useful to
color them, but this is a bit more dificult in this case because
the block title is language dependent and must be read from a table
header.