Greasemonkey user script added
Currently there is no add-on for firefox.
To mitigate it, a user script was created and it can be installed in
Greasemonkey.
It also serves as a proof of concept to be later used for the Add-On.

Currently there is a single user script that covers only the use
case to highlight the CI in shown in the note text.