Updated version and changed its semantic
Changed version to be compliant with Semantic Version 2.0 (semver.org).
Previous version was 0.0.7, even though the extensio is already being
used productively and that previous version changes were a mix of
new features and bug fixes.
New version is 1.8.0, in format X.Y.Z. X=1 because it is used
productively and Y = 8 as a continuation of what previously was Z = 7.