Add link to pre requisite CIs in CI headers
In order to create this link an addional parameter called pakid
is needed.
We take this parameter from the existing link to the current
CI(s) (first found is used).
This seems to be the only way to get this parameter.
It should be accurate in most cases, as it seems to be related to
the Software Component.
It may still fail in notes containing mixed Software Components.