feat: #28 Code highlighting in me.sap.com
This commit adds support for highlighting code in correction
instructions on me.sap.com.

It has been observed that such support was already partly
present as in some cases the CIs would be shown from
iframes that were loading from launchpad.support.sap.com,
but in some other cases, it is shown directly on the me.sap.com
page.

We observed that it worked when the url followed the pattern
https://me.sap.com/corrins/**, but not when it followed the
pattern https://me.sap.com/correctioninstruction/**

examples:
- https://me.sap.com/corrins/0003171185/15841
- https://me.sap.com/correctioninstruction/0003152012/41/corrinsdetail/0001116049/41/0020751258/0001/756/756/SAP_BASIS/Code

Related issues:
- https://github.com/dellagustin/SAP_Note_Enhancer/issues/28